+++ src/commands/exportAsPDF.ts
@@ -1,23 +1,20 @@
 import * as vscode from 'vscode'
 import * as fs from 'fs'
 import * as path from 'path'
-import { exec } from "child_process"
-import * as zlib from 'zlib';
+import { exec, spawn } from 'child_process'
+import extract = require('extract-zip')
+import fsExtra = require('fs-extra')
 import { https } from 'follow-redirects'
 import { isNullOrUndefined } from 'util'
-import { spawn } from "child_process";
-import { AsciidocParser } from '../text-parser';
-import { Command } from '../commandManager';
-import { AsciidocEngine } from '../asciidocEngine';
-import * as tmp from "tmp";
+import url = require('url')
 
-import HttpsProxyAgent = require('https-proxy-agent');
-import url = require('url');
-import { Logger } from '../logger';
+import { AsciidocEngine } from '../asciidocEngine'
+import { Command } from '../commandManager'
+import { Logger } from '../logger'
 
 export class ExportAsPDF implements Command
 {
-  public readonly id = 'asciidoc.exportAsPDF';
+  public readonly id = 'asciidoc.exportAsPDF'
 
   constructor(
     private readonly engine: AsciidocEngine,
@@ -58,17 +55,17 @@ export class ExportAsPDF implements Command
         .split(/(\s+)/)
         .filter(function (e) { return e.trim().length > 0 })
 
-      var adocpdf_cmd = adocpdf_cmd_array[0]
+      let adocpdf_cmd = adocpdf_cmd_array[0]
 
-      var adocpdf_cmd_args = adocpdf_cmd_array.slice(1)
+      let adocpdf_cmd_args = adocpdf_cmd_array.slice(1)
       adocpdf_cmd_args.push.apply(adocpdf_cmd_args, ['-q',
         '-B', '"' + docPath.dir.replace('"', '\\"') + '"',
         '-o', '"' + pdfPath.replace('"', '\\"') + '"', '-',
       ])
 
-      var options = { shell: true, cwd: docPath.dir }
+      let options = { shell: true, cwd: docPath.dir }
 
-      var asciidoctorpdf = spawn(adocpdf_cmd, adocpdf_cmd_args, options)
+      let asciidoctorpdf = spawn(adocpdf_cmd, adocpdf_cmd_args, options)
 
       asciidoctorpdf.stderr.on('data', (data) =>
       {
@@ -92,108 +89,83 @@ export class ExportAsPDF implements Command
     }
     else
     {
-      let wkhtmltopdf_path = vscode.workspace
+      let asciidoctorWebPdfPath = vscode.workspace
         .getConfiguration('asciidoc')
-        .get('wkhtmltopdf_path', '');
+        .get('asciidoctorWebPdfPath', '');
 
-      let parser = new AsciidocParser(path.resolve(doc.fileName))
-      //const body =  await parser.parseText()
-      const body = await this.engine.render(doc.uri, true, text, false, 'html5')
-      const ext_path = vscode.extensions.getExtension('asciidoctor.asciidoctor-vscode').extensionPath;
-      const html = body;
-      const showtitlepage = parser.getAttribute("showtitlepage")
-      const author = parser.getAttribute("author")
-      const email = parser.getAttribute("email")
-      const doctitle: string | undefined = parser.getAttribute("doctitle");
-      const titlepagelogo: string | undefined = parser.getAttribute("titlepagelogo");
-      const footer_center: string | undefined = parser.getAttribute("footer-center");
-      let cover: string | undefined = undefined;
-      let img_html: string = '';
-      if (!isNullOrUndefined(showtitlepage))
-      {
-        if (!isNullOrUndefined(titlepagelogo))
-        {
-          const image_url = titlepagelogo.startsWith('http') ? titlepagelogo : path.join(source_name.dir, titlepagelogo)
-          img_html = isNullOrUndefined(titlepagelogo) ? "" : `<img src="${image_url}">`
-        }
-        var tmpobj = tmp.fileSync({ postfix: '.html' });
-        let html = `\
-                <!DOCTYPE html>
-                <html>
-                    <head>
-                    <meta charset="UTF-8">
-                    <link rel="stylesheet" type="text/css" href="${ext_path + "/media/all-centered.css"}">
-                    </head>
-                    <body>
-                    <div class="outer">
-                        <div class="middle">
-                            <div class="inner">
-                                ${img_html}
-                                <h1>${doctitle}</h1>
-                                <p>${author} &lt;${email}&gt;</p>
-                            </div>
-                        </div>
-                    </div>
-                    </body>
-                </html>
-                `;
-        fs.writeFileSync(tmpobj.name, html, 'utf-8')
-        cover = `cover ${tmpobj.name}`;
-      }
+      const extPath = vscode.extensions.getExtension('asciidoctor.asciidoctor-vscode').extensionPath;
       const platform = process.platform
       const ext = platform == "win32" ? '.exe' : ''
-      const arch = process.arch;
-      var binary_path = path.resolve(path.join(__dirname, 'wkhtmltopdf-' + platform + '-' + arch + ext))
+      let binaryDir = path.resolve(path.join(extPath, 'bin', 'asciidoctor-web-pdf-' + platform))
+      let binaryPath = path.resolve(binaryDir, 'asciidoctor-web-pdf' + ext)
 
-      if (wkhtmltopdf_path != '')
-        binary_path = wkhtmltopdf_path;
+      if (asciidoctorWebPdfPath != '')
+        binaryPath = asciidoctorWebPdfPath;
 
-      if (!fs.existsSync(binary_path))
+      if (!fs.existsSync(binaryDir))
       {
-        var label = await vscode.window.showInformationMessage("This feature requires wkhtmltopdf\ndo you want to download", "Download")
+        let label = await vscode.window.showInformationMessage("This feature requires asciidoctor-web-pdf.\nDo you want to download?\nWarning ~200 Mb download", "Download")
         if (label != "Download")
           return
-        var error_msg = null
+          let error_msg = null
 
         await vscode.window.withProgress({
           location: vscode.ProgressLocation.Window,
-          title: "Downloading wkhtmltopdf",
+          title: "Downloading asciidoctor-web-pdf",
           // cancellable: true
         }, async (progress) =>
         {
-          progress.report({ message: 'Downloading wkhtmltopdf...' });
-          const download_url = `https://github.com/joaompinto/wkhtmltopdf/releases/download/v0.0.1/wkhtmltopdf-${platform}-${arch}${ext}.gz`
-          this.logger.log("Downloading " + download_url)
-          await download_file(download_url, binary_path + ".gz", progress).then(() =>
-          {
-            progress.report({ message: 'Unzipping wkhtmltopdf...' })
-            const ungzip = zlib.createGunzip()
-            const inp = fs.createReadStream(binary_path + ".gz")
-            const out = fs.createWriteStream(binary_path)
-            inp.pipe(ungzip).pipe(out)
-            fs.chmodSync(binary_path, 0x755)
+          progress.report({ message: 'Downloading asciidoctor-web-pdf...' });
+
+          const platformMapping = {
+            'win32': 'win',
+            'darwin': 'mac',
+            'linux': 'linux',
+          }
+          const asciidoctor_web_pdf_version = 'v1.0.0-alpha.12'
+          const download_url = `https://github.com/Mogztter/asciidoctor-web-pdf/releases/download/${asciidoctor_web_pdf_version}/asciidoctor-web-pdf-${platformMapping[platform]}-v1.0.0-alpha.12.zip`
+
+          this.logger.log("downloading asciidoctor-web-pdf from:" + download_url)
+          fs.mkdirSync(path.join(binaryDir,'../'))
+          fs.mkdirSync(binaryDir)
+          let filePath = path.join(binaryDir, 'asciidoctor-web-pdf.zip')
+          // TODO: Fixme path
+          await download_file(download_url, filePath, progress).then( async () => {
+            progress.report({ message: 'Unzipping asciidoctor-web-pdf...' })
+            await extract(filePath, {dir: binaryDir})
+            fs.unlinkSync(filePath)
+
           }).catch(async (reason) =>
           {
-            binary_path = null;
-            console.error("Error downloading", download_url, " ", reason)
-            await vscode.window.showErrorMessage("Error installing wkhtmltopdf, " + reason.toString())
+            console.error("Error downloading and extracting", download_url, " ", reason)
+            // we leave it clean in case of errors to allow a non-conflicting retry
+            fsExtra.removeSync(binaryDir)
+            binaryDir = null;
+            await vscode.window.showErrorMessage("Error installing asciidoctor-web-pdf, " + reason.toString())
             return
           })
         })
-        if (isNullOrUndefined(binary_path))
+        if (isNullOrUndefined(binaryDir))
           return;
       }
       var save_filename = await vscode.window.showSaveDialog({ defaultUri: pdf_filename })
-      if (!isNullOrUndefined(save_filename))
-      {
-        await html2pdf(html, binary_path, cover, footer_center, save_filename.fsPath)
+      
+      if (!isNullOrUndefined(save_filename)) {
+        await vscode.window.withProgress({
+          location: vscode.ProgressLocation.Window,
+          title: "Converting file",
+        }, async (progress) => {
+          progress.report({ message: 'Converting file...' })
+          await adoc2pdf(text, binaryPath, undefined, undefined, save_filename.fsPath)
           .then((result) => { offer_open(result) })
           .catch((reason) =>
           {
             console.error("Got error", reason)
-            vscode.window.showErrorMessage("Error converting to PDF, " + reason.toString());
+            vscode.window.showErrorMessage("Error converting to PDF, " + reason.toString())
           })
+        })
       }
+
     }
   }
 }
@@ -203,9 +175,9 @@ async function download_file(download_url: string, filename: string, progress)
 
   return new Promise((resolve, reject) =>
   {
-    var download_options = url.parse(download_url);
+    var download_options = url.parse(download_url)
     var wstream = fs.createWriteStream(filename)
-    var totalDownloaded = 0;
+    var totalDownloaded = 0
     // Proxy support needs to be reworked
     // var proxy = process.env.http_proxy || vscode.workspace.getConfiguration("http", null)["proxy"].trim();
     // var proxyStrictSSL = vscode.workspace.getConfiguration("http", null)["proxyStrictSSL"];
@@ -225,33 +197,31 @@ async function download_file(download_url: string, filename: string, progress)
         return reject("http error" + resp.statusCode)
       }
 
-      // A chunk of data has been recieved.
+      // A chunk of data has been received
       resp.on('data', (chunk) =>
       {
         totalDownloaded += chunk.length
-        progress.report({ message: "Downloading wkhtmltopdf ... " + ((totalDownloaded / contentSize) * 100.).toFixed(0) + "%" })
+        progress.report({ message: "Downloading asciidoctor-web-pdf... " + ((totalDownloaded / contentSize) * 100.).toFixed(0) + "%" })
         wstream.write(chunk)
       });
 
-      // The whole response has been received. Print out the result.
+      // The whole response has been received.
       resp.on('end', () =>
       {
         wstream.end()
-        resolve()
+        resolve(undefined)
       });
 
     }).on("error", (err) =>
     {
       console.error("Error: " + err.message);
       reject(err.message)
-    });
+    })
   })
 }
 
 function offer_open(destination)
 {
-
-  // Saving the JSON that represents the document to a temporary JSON-file.
   vscode.window.showInformationMessage(("Successfully converted to " + path.basename(destination)), "Open File").then((label: string) =>
   {
     if (label == "Open File")
@@ -263,45 +233,39 @@ function offer_open(destination)
         // breaks if there is a single quote (') in the path
         case 'win32':
           exec(`"${destination.replace('"', '\\"')}"`);
-          break;
+          break
         case 'darwin':
           exec(`\`open "${destination.replace('"', '\\"')}" ; exit\``);
-          break;
+          break
         case 'linux':
           exec(`\`xdg-open "${destination.replace('"', '\\"')}" ; exit\``);
-          break;
+          break
         default:
           vscode.window.showWarningMessage("Output type is not supported");
-          break;
+          break
       }
     }
   })
 }
 
-export async function html2pdf(html: string, binary_path: string, cover: string, footer_center: string, filename: string)
+export async function adoc2pdf(text: string, binary_path: string, cover: string, footer_center: string, filename: string)
 {
-  let documentPath = path.dirname(filename);
+  let documentPath = path.dirname(filename)
 
   return new Promise((resolve, reject) =>
   {
-    var options = { cwdir: documentPath, stdio: ['pipe', 'ignore', "pipe"] }
-    let cmd_arguments = ['--encoding', ' utf-8', '--javascript-delay', '1000'];
-    if (!isNullOrUndefined(footer_center))
-    {
-      cmd_arguments = cmd_arguments.concat(['--footer-center', footer_center])
-    }
-    if (!isNullOrUndefined(cover))
-    {
-      cmd_arguments = cmd_arguments.concat(cover.split(" "))
-    }
-    cmd_arguments = cmd_arguments.concat(['-', filename]);
-    var command = spawn(binary_path, cmd_arguments, options)
-    var error_data = '';
-    command.stdin.write(html);
-    command.stdin.end();
+    let options = { cwdir: documentPath, stdio: ['pipe', 'ignore', "pipe"] }
+    const krokiPath = path.dirname(require.resolve("asciidoctor-kroki/package.json"))
+    let cmd_arguments = ['-r', krokiPath, '-o', filename, '-']
+
+    let command = spawn(binary_path, cmd_arguments, options)
+    let error_data = ''
+
+    command.stdin.write(text)
+    command.stdin.end()
     command.stderr.on('data', (data) =>
     {
-      error_data += data;
+      error_data += data
     })
     command.on('close', (code) =>
     {
@@ -310,5 +274,5 @@ export async function html2pdf(html: string, binary_path: string, cover: string,
       else
         reject(error_data)
     })
-  });
+  })
 }